add filter feature

+++ lib/home/home.dart
@@ -26,6 +26,9 @@ class _HomeState extends State<Home> {
   final _streamController = StreamController<(DocumentChangeType, Product)>();
 
   bool? _isSortedByCategory;
+  int _filterCategory = 0;
+
+  Map<String, Product>? _holdCurrentState;
 
   @override
   void initState() {
@@ -38,17 +41,46 @@ class _HomeState extends State<Home> {
     WidgetsBinding.instance.addPostFrameCallback((_) {
       _streamController.stream.listen((event) {
         _isSortedByCategory = null;
-        switch (event.$1) {
-          case DocumentChangeType.added:
-          case DocumentChangeType.modified:
-            context.read<ProductCubit>().addOrUpdateIfExist(event.$2);
-            break;
-          case DocumentChangeType.removed:
-            // Support remove useless img on Firestorage
-            FirestorageService.instance.delete(event.$2.imgUrl ?? "");
-            context.read<ProductCubit>().remove(event.$2);
-            break;
-          default:
+        if (_holdCurrentState == null) {
+          switch (event.$1) {
+            case DocumentChangeType.added:
+            case DocumentChangeType.modified:
+              context.read<ProductCubit>().addOrUpdateIfExist(event.$2);
+              break;
+            case DocumentChangeType.removed:
+              // Support remove useless img on Firestorage
+              FirestorageService.instance.delete(event.$2.imgUrl ?? "");
+              context.read<ProductCubit>().remove(event.$2);
+              break;
+            default:
+          }
+        } else {
+          if (event.$2.categoryId == _filterCategory) {
+            switch (event.$1) {
+              case DocumentChangeType.added:
+              case DocumentChangeType.modified:
+                context.read<ProductCubit>().addOrUpdateIfExist(event.$2);
+                break;
+              case DocumentChangeType.removed:
+                // Support remove useless img on Firestorage
+                FirestorageService.instance.delete(event.$2.imgUrl ?? "");
+                context.read<ProductCubit>().remove(event.$2);
+                break;
+              default:
+            }
+          }
+          switch (event.$1) {
+            case DocumentChangeType.added:
+            case DocumentChangeType.modified:
+              _holdCurrentState!.addAll({event.$2.id!: event.$2});
+              break;
+            case DocumentChangeType.removed:
+              // Support remove useless img on Firestorage
+              FirestorageService.instance.delete(event.$2.imgUrl ?? "");
+              _holdCurrentState!.remove(event.$2.id!);
+              break;
+            default:
+          }
         }
       });
     });
@@ -75,6 +107,55 @@ class _HomeState extends State<Home> {
                 icon: const Icon(
                   CupertinoIcons.search,
                 )),
+            PopupMenuButton<int>(
+              icon: const Icon(Icons.filter_alt_rounded),
+              onSelected: (value) {
+                if (value != _filterCategory) {
+                  _filterCategory = value;
+                  switch (_filterCategory) {
+                    case 0:
+                      if (_holdCurrentState != null) {
+                        context
+                            .read<ProductCubit>()
+                            .replaceState(_holdCurrentState!);
+                      }
+                      // _holdCurrentState!.clear();
+                      _holdCurrentState = null;
+                      break;
+
+                    default:
+                      _holdCurrentState ??=
+                          Map.of(context.read<ProductCubit>().currentState());
+                      Map<String, Product> mapFilter = {};
+                      _holdCurrentState!.forEach((key, value) {
+                        if (value.categoryId == _filterCategory) {
+                          mapFilter.addAll({key: value});
+                        }
+                      });
+                      context.read<ProductCubit>().replaceState(mapFilter);
+                      break;
+                  }
+                }
+              },
+              itemBuilder: (BuildContext context) => <PopupMenuEntry<int>>[
+                const PopupMenuItem<int>(
+                  value: 0,
+                  child: Text('Remove filter'),
+                ),
+                const PopupMenuItem<int>(
+                  value: 1,
+                  child: Text('Category 1'),
+                ),
+                const PopupMenuItem<int>(
+                  value: 2,
+                  child: Text('Category 2'),
+                ),
+                const PopupMenuItem<int>(
+                  value: 3,
+                  child: Text('Category 3'),
+                ),
+              ],
+            ),
             PopupMenuButton<bool>(
               icon: const Icon(Icons.sort),
               onSelected: (value) {
